+++ docs/releases.md
@@ -3,6 +3,16 @@
 This page will give you the change that are occuring when a new version has been published on pypi.
 The changes have been tracked starting version 0.1.0
 
+## version 0.3.2
+
+* Adding annotation methods
+  * getAnnotations
+  * getAnnotation
+  * createAnnotation
+  * updateAnnotation
+  * deleteAnnotation
+* Add a `getDateRange` method
+
 ## version 0.3.1
 
 * adding the `workspace` module that allow more dynamic and better presentation of the data. [Documentation](https://github.com/pitchmuc/adobe-analytics-api-2.0/blob/master/docs/workspace.md)